feat(appHeader): rename HeaderLeft to AppHeader
